Implement detection of user overrides with some cases
On Windows, user can override certain locale parameters under Advanced
dialog in the Country & Language control panel and we want to represent
these.

Windows has convenient flag, LOCALE_NOUSEROVERRIDE, for the
GetLocaleInfo(Ex) function that lets us compare the normal value with
the user preference, so we can encode it if the user preference differs
from the default.

Since I am not sitting in front of Windows right now, I have only
implemented 3 items I know for sure are overridable—decimal and
thousand separator and digits—and will add the rest when I review the
content of the dialog—and test whether the approach works in the first
place.