Add proper error type
It has just one variant, because the only kind of error that ever makes
it to the user is a parse error when trying to construct a LanguageRange
or add something to a Locale, but there is some chance it will be
extended in future, so created an extensible enum for it.

This is interface change, so bump version to 0.2.0-pre too.