Reduce compile time and transitive deps
locale_config relies on regular expressions to parse input strings. The
expressions are quite complex, but the input data is tiny. Furthermore,
dealing with locales is unlikely to be in a hot loop. This lets us
disable performance optimizations of the "regex" crate, which in turn
drops some transitive dependencies and decreases the compile time of
locale_config.

With these changes, `cargo +nightly build -Z timings --release` went
down from 9.6 seconds to 5.8 seconds on my quad-core Intel i7-7700HQ.